Refactor bytes.go and bytes_test.go to smaller files, introduce go1.17 and go1.20 style of array copy
